requirements.txt
pyttsx3
python-vlc
